Use int64 instead of int
Otherwise we'd only be able to represent transactions
of up to 20k lol

+++ model/account.go
@@ -4,8 +4,8 @@ package model
 // as well as the account creation object representing its initial state. In a
 // multi-account setup, it should include some ID of the account.
 type Account struct {
-	ActiveCard     bool `json:"active-card"`
-	AvailableLimit int  `json:"available-limit"`
+	ActiveCard     bool  `json:"active-card"`
+	AvailableLimit int64 `json:"available-limit"`
 }
 
 func (a Account) Copy() *Account {
